can FINALLY reach post endpoint
Something about the parameters of post was messing it up, which in turn had to do with the deserialization of the object as i rely on newtonsoft. Even added use newtonsoft to program.cs which didnt fix the issue. Now relying on only sending pure strings to post methods which are the serialized objects.