Reached VG level!
Everything seems to be working as expected. As for the CustomUser that inherits from IdentityUser (just like applicationuser) it just has a string country property added to it. Just noticed i could have just extended applicationuser itself. Well, it still seems to be working so lets not do that. Also added classes to be used for conversion api. One for the api response, one to call the api and another static class to retrieve things related to countries, their three letter abbreviation and their currency abbreviation