Dialogbox is working correctly
Event creation is working 100%

Event deletion is working 100%

Dialog box shows up to confirm delete operation. If yes = delete, if no = nothing happens.

Need to resolve the column imageSrcs, it is not needed for the project but is not nullable and does not create event if left unassigned with a value.